pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "library-mcp"
version = "0.1.0"
description = "MCP server for semantic search over a personal ebook library"
requires-python = ">=3.11"
dependencies = [
    # Core
    "mcp[cli]",
    "chromadb",
    "sentence-transformers",
    "pydantic>=2.0",

    # Text extraction
    "pymupdf",
    "pdfplumber",
    "ebooklib",
    "mobi",
    "beautifulsoup4",
    "pytesseract",

    # CLI
    "typer",
    "rich",

    # Utilities
    "tiktoken",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
openai = ["openai"]
dev = ["pytest", "pytest-cov"]

[project.scripts]
library = "library_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/library_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
